# openapi.yaml
openapi: 3.1.0
info:
  title: Persistio API
  version: 0.1.19
  description: Self-hosted memory service for AI agents.
servers:
  - url: http://localhost:4827
paths:
...
  /v1/ingest:
    post:
      summary: Append raw conversation chunks
      description: |
        Appends raw conversation chunks for extraction. Each chunk must include
        the original conversation turn timestamp as an ISO 8601 datetime with
        UTC `Z` or an explicit offset. The
        live OpenClaw plugin supplies this automatically; custom clients and
        backfill tools must provide it explicitly.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, chunks]
              properties:
                session_id:
                  type: string
                chunks:
                  type: array
                  items:
                    type: object
                    required: [role, content, timestamp]
                    properties:
                      role:
                        type: string
                        enum: [user, assistant, tool]
                      content:
                        type: string
                        maxLength: 8000
                      timestamp:
                        type: string
                        format: date-time
                        description: Original conversation turn time, not ingest request time.
      responses:
        '202':
          description: Chunks accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: integer
                  chunks:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        created_at:
                          type: string
                          format: date-time
  /v1/ingest/bulk:
    post:
      summary: Bulk append raw conversation chunks
      description: |
        Premium backfill/import endpoint. Accepts up to 2048 chunks, writes raw
        chunks with one multi-row insert, and queues extraction at bulk priority.
        Request bodies are accepted up to `BULK_INGEST_BODY_LIMIT_BYTES`
        bytes, 25 MiB by default.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, chunks]
              properties:
                session_id:
                  type: string
                chunks:
                  type: array
                  maxItems: 2048
                  items:
                    type: object
                    required: [role, content, timestamp]
                    properties:
                      role:
                        type: string
                        enum: [user, assistant, tool]
                      content:
                        type: string
                        maxLength: 8000
                      timestamp:
                        type: string
                        format: date-time
                        description: Original conversation turn time, not ingest request time.
      responses:
        '202':
          description: Chunks accepted and extraction job queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: integer
                  job_id:
                    type: string
                    format: uuid
                  chunks:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        created_at:
                          type: string
                          format: date-time
        '403':
          description: Bulk ingest requires a premium plan
        '413':
          description: Too many chunks
...
  /v1/curation:
    get:
      summary: Fetch live Curator capacity, queue, and usage state
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Curator capacity status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurationStatus'
...
  /admin/vaults/{id}/curation:
    get:
      summary: Fetch staff Curator capacity, queue, and usage state for a vault
      security:
        - adminKey: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Curator capacity status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurationStatus'
        '404':
          description: Vault not found
components:
  schemas:
...
    CurationStatus:
      type: object
      properties:
        vault_id:
          type: string
          format: uuid
        plan:
          type: string
        period:
          type: string
          example: '2026-05'
        limits:
          type: object
          additionalProperties: true
        usage:
          type: object
          properties:
            curator_runs:
              type: integer
            curator_requests:
              type: integer
            curator_input_tokens:
              type: integer
            curator_output_tokens:
              type: integer
            curator_candidates_processed:
              type: integer
            curator_candidates_deferred:
              type: integer
        backlog:
          type: object
          properties:
            pending_queue_rows:
              type: integer
            oldest_queue_age_seconds:
              type: integer
              nullable: true
            backlog_limit:
              type: integer
            pressure:
              type: string
              enum: [ok, warning, over_limit]
        schedule:
          type: object
          properties:
            last_run_at:
              type: string
              nullable: true
            next_run_at:
              type: string
              nullable: true
            eligible_now:
              type: boolean
            defer_reason:
              type: string
              nullable: true
        recent_runs:
          type: array
          items:
            type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key